requirements.txt
python-telegram-bot==20.7
cloudscraper==1.2.71
beautifulsoup4==4.12.2
python-dotenv==1.0.0